move version to root
